Add husky pre-commit hook
Build and documentation NPM scripts will automatically run pre-commit. This is a better alternative to using GitHub Actions to commit changes as that would result in unnecessary double commits for every minor change.

The lint NPM script is not called as that is done by the GitHub Actions lint job.