Use Prettier over pretty
Prettier's HTML formatter offers more customizability and is better maintained than "pretty".

Documentation files generated with JSDoc are also now formatted.

Prettier uses the angular-html-parser for HTML files and as a result, threw an error when trying to format some generated documentation files. This was resolved by manually replacing all instances of the error with a fix.