Create `printMD`
If links printed to the canvas log are clicked normally, the URL is
opened in the canvas itself (which will result in a blank page except
for select domains). So, links must be opened indirectly.

The chosen method of accomplishing this is to use the `<abbr>` tag to
display a "Ctrl + Click" message on mouse hover and the `target="_blank"
property of the `<a>` tag to prevent links from being opened directly.